requirements.txt
asgiref==3.2.10
cachetools==4.2.2
certifi==2020.6.20
cffi==1.14.6
chardet==3.0.4
cloudinary==1.22.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==3.0
django-allauth==0.45.0
django-cors-headers==3.4.0
django-debug-toolbar==3.2.2
django-rest-passwordreset==1.1.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.7.2
drf-yasg==1.17.1
facebook-sdk==3.1.0
google-api-core==2.0.0
google-api-python-client==2.18.0
google-auth==2.0.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mysqlclient==2.0.1
numpy==1.20.2
oauthlib==3.1.1
packaging==21.0
pandas==1.2.3
paystackapi==2.0.0
Pillow==7.2.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
pyotp==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.19.0
python-memcached==1.59
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rest-condition==1.0.3
rsa==4.7.2
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
sentry-sdk==1.4.3
simplejson==3.17.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
